Move logging setup into app.serve
Otherwise, it would open the log file at app startup
This means that we cannot clean up ~/.ubuntupro/log because
the very same process is using it when running the 'clean' verb.
